Исправлен тип power с double на int в классе Car

+++ src/main/java/model/Car.java
@@ -90,7 +90,7 @@ public class Car implements Comparable<Car> {
             return this;
         }
 
-        public Builder setPower(double power) {
+        public Builder setPower(int power) {
             this.power = power;
             return this;
         }
